In `topics/topic_info` module, rename and change signature of the public trait `from_items` to `from_sal_topic` and implement trait.
This new trait receives a reference to `SalTopic`, a `topic_subname` string and an `indexed` boolean, that indicates if the component is indexed or not. It then generates the map of private fields and the topic fields as `FieldInfo`. Finally it return an instance of `TopicInfo` constructed with this information.