Initial implementation of read_topic.
Here we implement the majority of the synchronous methods, including how to push new data and pop data synchronously.
We also add some async code but the implementation is still very bare and does not perform what they are supposed to do, this will come later.

Note that the implementation wraps all the data that requires "interior mutability" with the standard (synchronous) version of `Arc<Mutec<>>`.
Because this code is intended to work with async, we should avoid locking too long on these resources.
Here we are only locking when pushing, popping or changing the data.
When in need to wait for data to arrive, we should probably attempt to do so using some async synchronization mechanism.