In `src/domain.rs`, add method `new_rc` that returns a `Domain` wrapped in `Rc<RefCell<Domain>>`.
In most conditions we need to wrap domain in a a reference counted (`Rc`) so that it can be shared with instances of `SalInfo`, which itself stores a weak reaference to the domain.
This new method helps creating the wrapped version, and provides some encapsulation in case we need to change this in the future.