In `src/sal_info`, add new method `add_reader` to `SalInfo`. This method receives a `ReadTopic` instance and adds it to the `read_topics` HashMap of `SalInfo`. This is required because the read loop is in `SalInfo`, which will contain all the topic read and write facility. When adding a "read topic" to `SalInfo` we basically make sure the read loop will handle data from that topic. Note that the read loop is still not implemented.
One big change required for this to work is the introduction of the lifetime `a` to `SalInfo` which, in turn, requires us to add the lifetime to several other structs, which are handled in the following commits.

Also note that the `write_data` method now has a different lifetime (`b`), because the data passed to `write_data` by the `WriteTopic` class is created inside the write method and has a different lifetime.