Add new executable ``set_summary_state``.
This program works similarly to the ``set_summary_state`` method in the original salobj library, but this is a program one can execute in the command line.
The program accepts a name, index (optional) and desired state.
When executed it will get the current state of the CSC and execute the required commands to leave the CSC in the desired state.