In ``src/topics/field_info.rs``, fix serialization of fields with array values.
While implementing the integration with the schema register I realized that the schema produced by serializing ``FieldInfo`` into avro was wrong.
Basically all fields were serialized as scalars.
In order to fix this I had to first implement my own Serialization trait for ``AvroSchemaField``.
Then ``FieldInfo`` struct and implementation were modified to be able to handle the scalar/array condition.
In the end, everything is treated as an array but when the field is defined as scalar in the schema it will serialize to a scalar value instead of a single value array.

Finally I also replaced HashMap with BTreeMap in ``AvroSchemaField`` to make sure the field order is preserved, which is somewhat important for the schema.
Basically, I noticed that if you change the order of the fields in the avro schema it results in different schema, even though they are the same.