In ``topics/topic_info.rs``, add new ``from_generic_sal_topic`` trait to ``TopicInfo``, that generates ``TopicInfo`` for generic topics.
Generic topics are named ``SALGeneric_<topic_type>_<topic_name>`` instead of ``<component_name>_<topic_type>_<topic_name>``.
Therefore, when generating ``TopicInfo`` for generic topics we need to pass in the name of the component instead of the name derived from the topic schema.
This method is designed to handle this for upstream modules, e.g. ``SALSubsystemInfo``.