In ``remote.rs``, update ``Remote.new`` trait to register topics before creating the topics objects.
In general this is necessary when the cluster has not been initialized by the system yet.
But the ``Domain.register_topics`` method used in the process will return immediately if the topics have already been loaded, so it is not such a burden to call it every time a remote is created.