In ``component_info``,  implement topic naming sanitation.
This module is problably the one with the largest code changes.

To start with, instead of storing ``TopicInfo`` using the topic sal name, use topic name instead.
This is a much cleaner implementation as we don't have to use the component name to get the ``TopicInfo`` for a topic.
This has consequences is several other modules though, but it is well worth the change.

In addition to the code changes, add documentation to the module.