In ``src/topics/read_topic.rs``, refactor ``ReadTopic`` to stop using unwrap and return ``SalObjResult`` in methods that can fail.
Also, note that now ``ReadTopic``, holds a ``KafkaResult<Consumer>`` instead of a ``Consumer``.
This allows one to postpone the creation of the consumer in case it fails and create an instance of ``ReadTopic`` even if there is no broker available.
The instance won't be able to read any data but it can still do some basic functionality and, in production, it gives the option to try to create the consumer after initialization.